agent/requirements.txt
# DineAI Voice Agent Requirements (Python)
# This file lists all Python packages needed for the voice agent

# LiveKit Agents Framework
livekit-agents[openai,deepgram,cartesia]>=1.0.0

# Additional dependencies
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
